Implement message encryption
What: Added classes to perform the
encryption of the message by taking
each word of the message and
performing some transformations on
the chars.
The Message handling is independent of
the encryption logic and the underlying
data-structure for the chars
Also, the encryption logic is kind of
idndependent from the data-structure
implementation.
Why: So any message can be encrypted,
and to keep modular code.